Move TensorData to its own file
